backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "aiomysql>=0.2.0",
    "alembic>=1.17.2",
    "fastapi>=0.116.1",
    "passlib[bcrypt]>=1.7.4",
    "pydantic-settings>=2.10.1",
    "python-dateutil>=2.9.0.post0",
    "python-dotenv>=1.1.1",
    "python-jose[cryptography]>=3.5.0",
    "python-multipart>=0.0.20",
    "sqlmodel>=0.0.24",
    "uvicorn>=0.35.0",
]
